Moving fileshare code around to prepare for further changes
