+++ fileshare/fileshare.go
@@ -8,9 +8,8 @@ import (
 )
 
 const (
-	DirDepthLimit       = 5
-	TransferFileLimit   = 1000
-	historySizeMaxBytes = 4 * 1024 * 1024 // 4Mb is also gRPC message default limit
+	DirDepthLimit     = 5
+	TransferFileLimit = 1000
 )
 
 // Fileshare defines a set of operations that any type that wants to act as a fileshare service
