--- fileshare/history_drop.go
@@ -1,7 +1,0 @@
-//go:build drop
-
-package fileshare
-
-func FileshareHistoryImplementation(storagePath string) Storage {
-	return NewJsonFile(storagePath)
-}
