--- fileshare/history_linux.go
@@ -1,7 +1,0 @@
-//go:build !drop
-
-package fileshare
-
-func FileshareHistoryImplementation(_ string) Storage {
-	return NoopStorage{}
-}
